refactor: add Preview and replace Upload.Selector
