+++ packages/core/client/src/collection-manager/templates/file.tsx
@@ -105,7 +105,7 @@ export const file: ICollectionTemplate = {
         uiSchema: {
           type: 'string',
           title: '{{t("Preview")}}',
-          'x-component': 'Upload.Selector',
+          'x-component': 'Preview.Selector',
           'x-read-pretty': true,
         },
       },
