+++ packages/core/client/src/schema-component/antd/index.ts
@@ -1,4 +1,5 @@
 export * from './action';
+export * from './association-select';
 export * from './block-item';
 export * from './calendar';
 export * from './card-item';
@@ -24,12 +25,12 @@ export * from './page';
 export * from './pagination';
 export * from './password';
 export * from './percent';
+export * from './preview';
 export * from './radio';
 export * from './record-picker';
+export * from './remote-select';
 export * from './rich-text';
 export * from './select';
-export * from './remote-select';
-export * from './association-select';
 export * from './space';
 export * from './table';
 export * from './table-v2';
