+++ packages/core/client/src/schema-component/antd/preview/index.md
@@ -1,0 +1,18 @@
+---
+nav:
+  path: /client
+group:
+  path: /schema-components
+---
+
+# Preview
+
+## Examples
+
+### Preview
+
+<code src="./demos/demo1.tsx" />
+
+### Preview
+
+<code src="./demos/demo2.tsx" />
