+++ packages/core/client/src/schema-component/antd/preview/index.ts
@@ -1,0 +1,1 @@
+export * from './Preview';
